+++ lib/widgets/lecture_simple_block.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:otlplus/constants/color.dart';
 import 'package:otlplus/constants/text_styles.dart';
 import 'package:otlplus/models/lecture.dart';
-import 'package:otlplus/utils/responsive_button.dart';
+import 'package:otlplus/widgets/responsive_button.dart';
 
 class LectureSimpleBlock extends StatelessWidget {
   final Lecture lecture;
