+++ lib/widgets/search_filter_panel.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:otlplus/constants/color.dart';
 import 'package:otlplus/constants/text_styles.dart';
 import 'package:otlplus/models/filter.dart';
-import 'package:otlplus/utils/responsive_button.dart';
+import 'package:otlplus/widgets/responsive_button.dart';
 
 class SearchFilterPanel extends StatefulWidget {
   final Map<String, FilterGroupInfo> filter;
