+++ lib/widgets/timetable_block.dart
@@ -5,7 +5,7 @@ import 'package:otlplus/constants/text_styles.dart';
 import 'package:otlplus/extensions/lecture.dart';
 import 'package:otlplus/models/lecture.dart';
 import 'package:otlplus/utils/get_text_height.dart';
-import 'package:otlplus/utils/responsive_button.dart';
+import 'package:otlplus/widgets/responsive_button.dart';
 
 class TimetableBlock extends StatelessWidget {
   final Lecture lecture;
